reorganize store code and add tests of root actions
