+++ package.json
@@ -30,7 +30,9 @@
     "ld-navigation": "^0.5.1",
     "rdf-ext": "^1.1.2",
     "rdf-fetch": "^1.0.0",
-    "vuex": "^3.1.0"
+    "vuex": "^3.1.0",
+    "readable-stream": "latest",
+    "@rdfjs/parser-jsonld": "latest"
   },
   "devDependencies": {
     "@babel/core": "^7.3.3",
