+++ src/store/root/action-types.js
@@ -1,0 +1,2 @@
+export const LOAD_RESOURCE = 'loadResource'
+export const SAVE_RESOURCE = 'saveResource'
