+++ src/store/root/getters.js
@@ -1,0 +1,14 @@
+import LdNavigator from 'ld-navigation/LdNavigator'
+
+export default {
+  resources (state) {
+    if (!state.resourceGraph) {
+      return []
+    }
+
+    return state.resourceGraph['@graph'] || []
+  },
+  resourceIri () {
+    return () => LdNavigator.resourceUrl
+  }
+}
