+++ src/store/root/mutation-types.js
@@ -1,0 +1,1 @@
+export const RESOURCE_LOADED = 'resource loaded'
