+++ src/store/root/mutations.js
@@ -1,0 +1,7 @@
+import { RESOURCE_LOADED } from './mutation-types'
+
+export default {
+  [RESOURCE_LOADED] (state, graphJson) {
+    state.resourceGraph = graphJson
+  }
+}
