+++ test/store/pipeline-actions.spec.js
@@ -2,334 +2,337 @@ import { expect, assert } from 'chai'
 import {
   load, save, addStep, saveVariable,
   deleteStep, selectStep, updateStep,
-  addVariable, deleteVariable } from '../../src/store/pipeline-action-types'
-import * as mutations from '../../src/store/pipeline-mutation-types'
-import actions from '../../src/store/pipeline-actions'
+  addVariable, deleteVariable } from '../../src/store/pipeline/action-types'
+import * as mutations from '../../src/store/pipeline/mutation-types'
+import actions from '../../src/store/pipeline/actions'
 import * as sinon from 'sinon'
-import Client from '../../src/Client'
-import cf from 'clownface'
-import rdf from 'rdf-ext'
 import { code } from '../../src/utils/namespaces'
+import * as rootActions from '../../src/store/root/action-types'
 
-describe('action', () => {
-  describe(load, () => {
-    it('commits iri to state', async () => {
-      // given
-      const client = sinon.mock(new Client())
-      const commit = sinon.spy()
-      const rootState = { client: client.object }
-      const iri = 'urn:pipeline:1'
-      client.expects('fetch').once().resolves(cf(rdf.dataset()))
+describe('pipeline store', () => {
+  describe('action', () => {
+    describe(load, () => {
+      it('commits iri to state', async () => {
+        // given
+        const dispatch = sinon.spy()
+        const commit = sinon.spy()
+        const iri = 'urn:pipeline:1'
+        const rootGetters = {
+          resources: []
+        }
 
-      // when
-      await actions.load({ commit, rootState }, iri)
+        // when
+        await actions.load({ commit, dispatch, rootGetters }, iri)
 
-      // then
-      expect(commit.firstCall.args[0]).to.equal(mutations.IRI_SET)
+        // then
+        expect(commit.firstCall.args[0]).to.equal(mutations.IRI_SET)
+      })
+
+      it('commits the pipeline body to state', async () => {
+        // given
+        const dispatch = sinon.spy()
+        const commit = sinon.spy()
+        const iri = 'urn:pipeline:1'
+        const rootGetters = {
+          resources: [{ id: iri }]
+        }
+
+        // when
+        await actions.load({ commit, dispatch, rootGetters }, iri)
+
+        // then
+        assert(commit.calledWith(
+          mutations.PIPELINE_SELECTED,
+          sinon.match({ id: iri })
+        ))
+      })
     })
 
-    it('commits the pipeline body to state', async () => {
-      // given
-      const client = sinon.mock(new Client())
-      const commit = sinon.spy()
-      const rootState = { client: client.object }
-      const iri = 'urn:pipeline:1'
-      client.expects('fetch').once().resolves(cf(rdf.dataset()))
-
-      // when
-      await actions.load({ commit, rootState }, iri)
-
-      // then
-      expect(commit.secondCall.args[0]).to.equal(mutations.PIPELINE_LOADED)
-      expect(commit.secondCall.args[1]['@graph']).to.be.an('array')
-    })
-  })
-
-  describe(save, () => {
-    it('saves the pipeline', async () => {
-      // given
-      const state = {
-        iri: 'urn:test:pipeline',
-        '@graph': {
-          'urn:test:pipeline': {
-            '@type': code.Pipeline
+    describe(save, () => {
+      it('saves the root resource', async () => {
+        // given
+        const state = {
+          iri: 'urn:test:pipeline',
+          '@graph': {
+            'urn:test:pipeline': {
+              '@type': code.Pipeline
+            }
           }
         }
-      }
-      const dispatch = sinon.spy()
-      const client = sinon.mock(new Client())
-      const rootState = { client: client.object }
-      client.expects('update').once()
+        const dispatch = sinon.spy()
 
-      // when
-      await actions.save({ state, rootState, dispatch })
+        // when
+        await actions.save({ state, dispatch })
 
-      // then
-      client.verify()
-    })
+        // then
+        assert(dispatch.calledWith(
+          rootActions.SAVE_RESOURCE,
+          null,
+          sinon.match({ root: true })
+        ))
+      })
 
-    it('reloads the pipeline', async () => {
-      // given
-      const state = {
-        iri: 'urn:test:pipeline',
-        '@graph': {
-          'urn:test:pipeline': {
-            '@type': code.Pipeline
+      it('reloads the pipeline', async () => {
+        // given
+        const state = {
+          iri: 'urn:test:pipeline',
+          '@graph': {
+            'urn:test:pipeline': {
+              '@type': code.Pipeline
+            }
           }
         }
-      }
-      const dispatch = sinon.spy()
-      const client = sinon.mock(new Client())
-      const rootState = { client: client.object }
-      client.expects('update').once()
+        const dispatch = sinon.spy()
 
-      // when
-      await actions.save({ state, rootState, dispatch })
+        // when
+        await actions.save({ state, dispatch })
 
-      // then
-      expect(dispatch.firstCall.args).to.deep.equal([
-        load, state.iri
-      ])
-    })
-  })
-
-  describe(addStep, () => {
-    it('commits an empty step', () => {
-      // given
-      const commit = sinon.spy()
-      const getters = {}
-
-      // when
-      actions.addStep({ commit, getters }, 5)
-
-      // then
-      assert(commit.calledWith(
-        mutations.STEP_ADDED,
-        sinon.match({
-          index: 5,
-          step: sinon.match.has('id')
-            .and(sinon.match.has('code:implementedBy', sinon.match.object))
-            .and(sinon.match.has('code:arguments', sinon.match.array))
-        })
-      ))
+        // then
+        assert(dispatch.calledWith(
+          load,
+          state.iri)
+        )
+      })
     })
 
-    it('selects the new step step', () => {
-      // given
-      const commit = sinon.spy()
-      const getters = {}
+    describe(addStep, () => {
+      it('commits an empty step', () => {
+        // given
+        const commit = sinon.spy()
+        const getters = {}
 
-      // when
-      actions.addStep({ commit, getters }, 5)
+        // when
+        actions.addStep({ commit, getters }, 5)
 
-      // then
-      assert(commit.calledWith(
-        mutations.STEP_SELECTED
-      ))
-    })
-  })
+        // then
+        assert(commit.calledWith(
+          mutations.STEP_ADDED,
+          sinon.match({
+            index: 5,
+            step: sinon.match.has('id')
+              .and(sinon.match.has('code:implementedBy', sinon.match.object))
+              .and(sinon.match.has('code:arguments', sinon.match.array))
+          })
+        ))
+      })
 
-  describe(deleteStep, () => {
-    it('unselects steps', () => {
-      // given
-      const commit = sinon.spy()
-      const index = 3
+      it('selects the new step step', () => {
+        // given
+        const commit = sinon.spy()
+        const getters = {}
 
-      // when
-      actions.deleteStep({ commit }, index)
+        // when
+        actions.addStep({ commit, getters }, 5)
 
-      // then
-      assert(commit.calledWith(
-        mutations.STEP_REMOVED,
-        index
-      ))
-    })
-
-    it('unselects steps', () => {
-      // given
-      const commit = sinon.spy()
-      const index = 3
-
-      // when
-      actions.deleteStep({ commit }, index)
-
-      // then
-      assert(commit.calledWith(
-        mutations.STEP_SELECTED,
-        null
-      ))
-    })
-  })
-
-  describe(selectStep, () => {
-    it('commits the selection', () => {
-      // given
-      const commit = sinon.spy()
-      const step = {}
-
-      // when
-      actions.selectStep({ commit }, step)
-
-      // then
-      assert(commit.calledWith(
-        mutations.STEP_SELECTED,
-        step
-      ))
-    })
-  })
-
-  describe(updateStep, () => {
-    it('clears selection', () => {
-      // given
-      const commit = sinon.spy()
-      const getters = {
-        steps: []
-      }
-
-      // when
-      actions.updateStep({ commit, getters }, {})
-
-      // then
-      assert(commit.calledWith(
-        mutations.STEP_SELECTED,
-        null
-      ))
+        // then
+        assert(commit.calledWith(
+          mutations.STEP_SELECTED
+        ))
+      })
     })
 
-    it('commits updated step', () => {
-      // given
-      const commit = sinon.spy()
-      const getters = {
-        steps: [{
-          id: 'step1'
-        }, {
+    describe(deleteStep, () => {
+      it('unselects steps', () => {
+        // given
+        const commit = sinon.spy()
+        const index = 3
+
+        // when
+        actions.deleteStep({ commit }, index)
+
+        // then
+        assert(commit.calledWith(
+          mutations.STEP_REMOVED,
+          index
+        ))
+      })
+
+      it('unselects steps', () => {
+        // given
+        const commit = sinon.spy()
+        const index = 3
+
+        // when
+        actions.deleteStep({ commit }, index)
+
+        // then
+        assert(commit.calledWith(
+          mutations.STEP_SELECTED,
+          null
+        ))
+      })
+    })
+
+    describe(selectStep, () => {
+      it('commits the selection', () => {
+        // given
+        const commit = sinon.spy()
+        const step = {}
+
+        // when
+        actions.selectStep({ commit }, step)
+
+        // then
+        assert(commit.calledWith(
+          mutations.STEP_SELECTED,
+          step
+        ))
+      })
+    })
+
+    describe(updateStep, () => {
+      it('clears selection', () => {
+        // given
+        const commit = sinon.spy()
+        const getters = {
+          steps: []
+        }
+
+        // when
+        actions.updateStep({ commit, getters }, {})
+
+        // then
+        assert(commit.calledWith(
+          mutations.STEP_SELECTED,
+          null
+        ))
+      })
+
+      it('commits updated step', () => {
+        // given
+        const commit = sinon.spy()
+        const getters = {
+          steps: [{
+            id: 'step1'
+          }, {
+            id: 'step2'
+          }, {
+            id: 'step3'
+          }]
+        }
+        const step = {
           id: 'step2'
-        }, {
-          id: 'step3'
-        }]
-      }
-      const step = {
-        id: 'step2'
-      }
+        }
 
-      // when
-      actions.updateStep({ commit, getters }, step)
+        // when
+        actions.updateStep({ commit, getters }, step)
 
-      // then
-      assert(commit.calledWith(
-        mutations.STEP_UPDATED,
-        sinon.match.has('index', 1)
-          .and(sinon.match.has('step', step))
-      ))
+        // then
+        assert(commit.calledWith(
+          mutations.STEP_UPDATED,
+          sinon.match.has('index', 1)
+            .and(sinon.match.has('step', step))
+        ))
+      })
+
+      it('does not commit step if not found in state', () => {
+        // given
+        const commit = sinon.spy()
+        const getters = {
+          steps: [{
+            id: 'step1'
+          }, {
+            id: 'step2'
+          }, {
+            id: 'step3'
+          }]
+        }
+        const step = {
+          id: 'step4'
+        }
+
+        // when
+        actions.updateStep({ commit, getters }, step)
+
+        // then
+        assert(commit.neverCalledWith(
+          mutations.STEP_UPDATED
+        ))
+      })
     })
 
-    it('does not commit step if not found in state', () => {
-      // given
-      const commit = sinon.spy()
-      const getters = {
-        steps: [{
-          id: 'step1'
-        }, {
-          id: 'step2'
-        }, {
-          id: 'step3'
-        }]
-      }
-      const step = {
-        id: 'step4'
-      }
+    describe(addVariable, () => {
+      it('appends the variable', () => {
+        // given
+        const name = 'foo'
+        const value = 'bar'
+        const commit = sinon.spy()
+        const getters = {
+          variables: [ {}, {} ]
+        }
 
-      // when
-      actions.updateStep({ commit, getters }, step)
+        // when
+        actions.addVariable({ commit, getters }, { name, value })
 
-      // then
-      assert(commit.neverCalledWith(
-        mutations.STEP_UPDATED
-      ))
-    })
-  })
+        // then
+        assert(commit.calledWith(
+          mutations.REPLACE_VARIABLES,
+          [
+            sinon.match.object,
+            sinon.match.object,
+            sinon.match({ name, value })
+          ]
+        ))
+      })
 
-  describe(addVariable, () => {
-    it('appends the variable', () => {
-      // given
-      const name = 'foo'
-      const value = 'bar'
-      const commit = sinon.spy()
-      const getters = {
-        variables: [ {}, {} ]
-      }
+      it('ignores superfluous properties', () => {
+        // given
+        const name = 'foo'
+        const value = 'bar'
+        const commit = sinon.spy()
+        const getters = {
+          variables: [ ]
+        }
 
-      // when
-      actions.addVariable({ commit, getters }, { name, value })
+        // when
+        actions.addVariable({ commit, getters }, { name, value, something: 'else' })
 
-      // then
-      assert(commit.calledWith(
-        mutations.REPLACE_VARIABLES,
-        [
-          sinon.match.object,
-          sinon.match.object,
-          sinon.match({ name, value })
-        ]
-      ))
+        // then
+        expect(commit.firstCall.lastArg[0]).to.deep.equal({ name, value })
+      })
     })
 
-    it('ignores superfluous properties', () => {
-      // given
-      const name = 'foo'
-      const value = 'bar'
-      const commit = sinon.spy()
-      const getters = {
-        variables: [ ]
-      }
+    describe(deleteVariable, () => {
+      it('commits array with variable removed', () => {
+        // given
+        const one = {}
+        const two = {}
+        const three = {}
+        const getters = {
+          variables: [ one, two, three ]
+        }
+        const commit = sinon.spy()
 
-      // when
-      actions.addVariable({ commit, getters }, { name, value, something: 'else' })
+        // when
+        actions.deleteVariable({ commit, getters }, 1)
 
-      // then
-      expect(commit.firstCall.lastArg[0]).to.deep.equal({ name, value })
+        // then
+        assert(commit.calledWith(
+          mutations.REPLACE_VARIABLES,
+          [
+            one, three
+          ]
+        ))
+      })
     })
-  })
 
-  describe(deleteVariable, () => {
-    it('commits array with variable removed', () => {
-      // given
-      const one = {}
-      const two = {}
-      const three = {}
-      const getters = {
-        variables: [ one, two, three ]
-      }
-      const commit = sinon.spy()
+    describe(saveVariable, () => {
+      it('replaces the saved variable', () => {
+        // given
+        const commit = sinon.spy()
+        const getters = {
+          variables: [{ name: 'old', value: 'foo' }]
+        }
 
-      // when
-      actions.deleteVariable({ commit, getters }, 1)
+        // when
+        actions.saveVariable({ commit, getters }, { index: 0, name: 'new', value: 'bar' })
 
-      // then
-      assert(commit.calledWith(
-        mutations.REPLACE_VARIABLES,
-        [
-          one, three
-        ]
-      ))
-    })
-  })
-
-  describe(saveVariable, () => {
-    it('replaces the saved variable', () => {
-      // given
-      const commit = sinon.spy()
-      const getters = {
-        variables: [{ name: 'old', value: 'foo' }]
-      }
-
-      // when
-      actions.saveVariable({ commit, getters }, { index: 0, name: 'new', value: 'bar' })
-
-      // then
-      assert(commit.calledWith(
-        mutations.REPLACE_VARIABLES,
-        [ sinon.match({ name: 'new', value: 'bar' }) ]
-      ))
+        // then
+        assert(commit.calledWith(
+          mutations.REPLACE_VARIABLES,
+          [ sinon.match({ name: 'new', value: 'bar' }) ]
+        ))
+      })
     })
   })
 })
